+++ R/plot_point_sd.R
@@ -11,8 +11,6 @@
 #' `ColSeq` (logical TRUE/FALSE) decides whether colours are picked by respecting the order in the palette or the most distant ones using \code{\link[grDevices]{colorRampPalette}}.
 #' 
 #' If there are many groups along the X axis and you prefer a single colour for the graph,use the `SingleColour` argument.
-#' 
-#' You are instead encouraged to show all data using the following functions: \code{\link{plot_scatterbar_sd}}, \code{\link{plot_scatterbox}}, \code{\link{plot_dotbox}}, \code{\link{plot_dotbar_sd}}, \code{\link{plot_scatterviolin}} or \code{\link{plot_dotviolin}}.
 #'
 #' @param data a data table object, e.g. data.frame or tibble.
 #' @param xcol name of the column with a categorical X variable.
@@ -22,11 +20,10 @@
 #' @param symsize size of point symbols, default set to 3.5.
 #' @param s_alpha fractional opacity of symbols, default set to 1 (i.e. maximum opacity & zero transparency).
 #' @param symshape The mean is shown with symbol of the shape number 21 (default, filled circle). Pick a number between 0-25 to pick a different type of symbol from ggplot2.  
-#' @param all_alpha fractional opacity of all data points (default = 0; i.e., not shown). Set to non-zero value if you would like all data points plotted in addition to the mean.
+#' @param all_alpha fractional opacity of all data points (default = 0.3). Set to non-zero value if you would like all data points plotted in addition to the mean.
 #' @param all_size size of symbols of all data points, if shown (default = 2.5).
 #' @param all_jitter reduce overlap of all data points, if shown, by setting a value between 0-1 (default = 0).
 #' @param all_shape all data points are shown with symbols of the shape number 1 (default, transparent circle). Pick a number between 0-25 to pick a different type of symbol from ggplot2.
-#' @param ewid width of error bars, default set to 0.2.
 #' @param TextXAngle orientation of text on X-axis; default 0 degrees. Change to 45 or 90 to remove overlapping text.
 #' @param LogYTrans transform Y axis into "log10" or "log2"
 #' @param LogYBreaks argument for \code{ggplot2[scale_y_continuous]} for Y axis breaks on log scales, default is `waiver()`, or provide a vector of desired breaks.
@@ -35,6 +32,7 @@
 #' @param facet_scales whether or not to fix scales on X & Y axes for all facet facet graphs. Can be `fixed` (default), `free`, `free_y` or `free_x` (for Y and X axis one at a time, respectively).
 #' @param fontsize parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.
 #' @param symthick thickness of symbol border, default set to `fontsize`/22.
+#' @param ewid width of error bars, default set to 0.2.
 #' @param ethick thickness of error bar lines; default `fontsize`/22.
 #' @param ColPal grafify colour palette to apply, default "okabe_ito"; see \code{\link{graf_palettes}} for available palettes.
 #' @param ColSeq logical TRUE or FALSE. Default TRUE for sequential colours from chosen palette. Set to FALSE for distant colours, which will be applied using  \code{scale_fill_grafify2}.
@@ -51,12 +49,12 @@
 #' plot_point_sd(data = data_doubling_time, 
 #' xcol = Student, ycol = Doubling_time)
 #' 
-#' #show all data points
+#' #hide all data points
 #' plot_point_sd(data = data_2w_Tdeath, 
 #' xcol = Genotype, ycol = PI, 
-#' facet = Time, all_alpha = 0.4)
+#' facet = Time, all_alpha = 0)
 
-plot_point_sd <- function(data, xcol, ycol, facet, ErrorType = "SD", symsize = 3.5, s_alpha = 1, symshape = 21, all_alpha = 0, all_size = 2.5, all_shape = 1, all_jitter = 0, ewid = 0.2, TextXAngle = 0, LogYTrans, LogYBreaks = waiver(), LogYLabels = waiver(), LogYLimits = NULL, facet_scales = "fixed", fontsize = 20, symthick, ethick, ColPal = c("okabe_ito", "all_grafify", "bright",  "contrast",  "dark",  "fishy",  "kelly",  "light",  "muted",  "pale",  "r4",  "safe",  "vibrant"), ColSeq = TRUE, ColRev = FALSE, SingleColour = "NULL", ...){
+plot_point_sd <- function(data, xcol, ycol, facet, ErrorType = "SD", symsize = 3.5, s_alpha = 1, symshape = 21, all_alpha = 0.3, all_size = 2.5, all_shape = 1, all_jitter = 0, ewid = 0.2, TextXAngle = 0, LogYTrans, LogYBreaks = waiver(), LogYLabels = waiver(), LogYLimits = NULL, facet_scales = "fixed", fontsize = 20, symthick, ethick, ColPal = c("okabe_ito", "all_grafify", "bright",  "contrast",  "dark",  "fishy",  "kelly",  "light",  "muted",  "pale",  "r4",  "safe",  "vibrant"), ColSeq = TRUE, ColRev = FALSE, SingleColour = "NULL", ...){
   ColPal <- match.arg(ColPal)
   if (!(ErrorType %in% c("SD", "SEM", "CI95"))) {
     stop('ErrorType should be "SD", "SEM" or "CI95".')}
@@ -65,6 +63,8 @@ plot_point_sd <- function(data, xcol, ycol, facet, ErrorType = "SD", symsize = 3
   if(ErrorType == "CI95") {ER <- "mean_cl_normal"}
   if (missing(ethick)) {ethick = fontsize/22}
   if (missing(symthick)) {symthick = fontsize/22}
+  if(symshape < 21 | symshape > 25){
+    stop("`symshape` should be betwee 21-25.")}
   if (ER == "mean_cl_normal") {
     suppressWarnings(P <- ggplot2::ggplot(data, aes(x = factor({{ xcol }}),
                                   y = {{ ycol }}))+